+++ smallworld/analyses/__init__.py
@@ -1,12 +1,16 @@
 from .analysis import Analysis, Filter
 from .code_coverage import CodeCoverage
+from .code_reachable import CodeReachable
 from .colorizer import ColorizerAnalysis
+from .colorizer_summary import ColorizerSummary
 from .control_flow_tracer import ControlFlowTracer
 
 __all__ = [
     "Analysis",
     "Filter",
     "ColorizerAnalysis",
+    "ColorizerSummary",
     "ControlFlowTracer",
     "CodeCoverage",
+    "CodeReachable",
 ]