+++ smallworld/hinting.py
@@ -41,13 +41,24 @@ class ControlFlowHint(Hint):
 
 @dataclass(frozen=True)
 class CoverageHint(Hint):
-    """Hold the a map of program counter to hit counter for an execution
+    """Holds the a map of program counter to hit counter for an execution.
 
     Arguments:
         coverage: A map from program counter to hit count
     """
 
     coverage: typing.Dict[int, int]
+
+
+@dataclass(frozen=True)
+class ReachableCodeHint(Hint):
+    """Indicates that we can get to a given program counter with symbolic execution.
+
+    Arguments:
+        address: The address we can reach
+    """
+
+    address: int
 
 
 @dataclass(frozen=True)
@@ -336,6 +347,7 @@ class DynamicValueHint(Hint):
     micro_exec_num: int
     instruction_num: int
     dynamic_value: str
+    color: int
     size: int
     use: bool
     new: bool
@@ -391,6 +403,7 @@ class DynamicMemoryValueHint(DynamicValueHint):
 class DynamicValueProbHint(Hint):
     instruction: typing.Any
     pc: int
+    color: int
     size: int
     use: bool
     new: bool